Create a utils.rs file for helper methods
