+++ zaino-proto/src/proto.rs
@@ -3,3 +3,4 @@
 pub mod compact_formats;
 pub mod proposal;
 pub mod service;
+pub mod utils;
