+++ zaino-proto/src/proto/utils.rs
@@ -1,0 +1,42 @@
+use crate::proto::service::PoolType;
+
+/// Errors that can arise when mapping `PoolType` from an `i32` value.
+pub enum PoolTypeError {
+    /// Pool Type value was map to the enum `PoolType::Invalid`.
+    InvalidPoolType,
+    /// Pool Type value was mapped to value that can't be mapped to a known pool type.
+    UnknownPoolType(i32),
+}
+
+// Converts a vector of pool_types (i32) into its rich-type representation
+// Returns `None` when invalid `pool_types` are found
+pub fn pool_types_from_vector(pool_types: &[i32]) -> Result<Vec<PoolType>, PoolTypeError> {
+    let pools = if pool_types.is_empty() {
+        vec![PoolType::Sapling, PoolType::Orchard]
+    } else {
+        let mut pools: Vec<PoolType> = vec![];
+
+        for pool in pool_types.iter() {
+            match PoolType::try_from(*pool) {
+                Ok(pool_type) => {
+                    if pool_type == PoolType::Invalid {
+                        return Err(PoolTypeError::InvalidPoolType);
+                    } else {
+                        pools.push(pool_type);
+                    }
+                }
+                Err(_) => {
+                    return Err(PoolTypeError::UnknownPoolType(*pool));
+                }
+            };
+        }
+
+        pools.clone()
+    };
+    Ok(pools)
+}
+
+/// Converts a `Vec<Pooltype>` into a `Vec<i32>`
+pub fn pool_types_into_i32_vec(pool_types: Vec<PoolType>) -> Vec<i32> {
+    pool_types.iter().map(|p| *p as i32).collect()
+}
