+++ zaino-state/src/backends/state.rs
@@ -47,6 +47,9 @@ use zaino_proto::proto::{
         GetAddressUtxosReply, GetAddressUtxosReplyList, LightdInfo, PingResponse, PoolType,
         RawTransaction, SendResponse, TransparentAddressBlockFilter, TreeState, TxFilter,
     },
+    utils::{
+        pool_types_from_vector, PoolTypeError
+    }
 };
 
 use zcash_protocol::consensus::NetworkType;
@@ -1037,42 +1040,6 @@ impl StateServiceSubscriber {
         times.sort_unstable();
         Ok(times[times.len() / 2])
     }
-}
-
-/// Errors that can arise when mapping `PoolType` from an `i32` value.
-enum PoolTypeError {
-    /// Pool Type value was map to the enum `PoolType::Invalid`.
-    InvalidPoolType,
-    /// Pool Type value was mapped to value that can't be mapped to a known pool type.
-    UnknownPoolType(i32),
-}
-
-// Converts a vector of pool_types (i32) into its rich-type representation
-// Returns `None` when invalid `pool_types` are found
-fn pool_types_from_vector(pool_types: &[i32]) -> Result<Vec<PoolType>, PoolTypeError> {
-    let pools = if pool_types.is_empty() {
-        vec![PoolType::Sapling, PoolType::Orchard]
-    } else {
-        let mut pools: Vec<PoolType> = vec![];
-
-        for pool in pool_types.iter() {
-            match PoolType::try_from(*pool) {
-                Ok(pool_type) => {
-                    if pool_type == PoolType::Invalid {
-                        return Err(PoolTypeError::InvalidPoolType);
-                    } else {
-                        pools.push(pool_type);
-                    }
-                }
-                Err(_) => {
-                    return Err(PoolTypeError::UnknownPoolType(*pool));
-                }
-            };
-        }
-
-        pools.clone()
-    };
-    Ok(pools)
 }
 
 #[async_trait]
